requirements.txt
# Core dependencies
pymongo==4.6.1
pyyaml==6.0.1
requests==2.31.0
python-dotenv==1.0.0

# AI/ML models
faster-whisper==1.1.1
transformers==4.36.2
torch==2.1.2
sentencepiece==0.1.99
protobuf==4.25.1
sacremoses==0.1.1

# Utilities
tqdm==4.66.1
ffmpeg-python==0.2.0

# Logging
coloredlogs==15.0.1

# Hive blockchain
beem
